+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## v0.1.0 (2023-11-16)
 
+- Develop `ExXendit.Transaction.get`
 - Develop `ExXendit.Transaction.list`
 - Initial release